requirements.txt
anthropic>=0.40.0
playwright>=1.40.0
pydantic>=2.0.0
jinja2>=3.1.0
aiofiles>=23.0.0
rich>=13.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
